requirements.txt
Flask==3.1.1
Flask-CORS==6.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.41
python-dotenv==1.1.1
openai==1.91.0
gunicorn==23.0.0
psycopg2-binary==2.9.10

